+++ src/common/spreadsheet/ASpreadSheetCell.hpp
@@ -15,7 +15,7 @@
 // General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// along with GNU Emacs.  If not, see <http://www.gnu.org/licenses/>.
+// along with SimTaDyn.  If not, see <http://www.gnu.org/licenses/>.
 //=====================================================================
 
 #ifndef ASPREADSHEETCELL_HPP_
@@ -159,7 +159,7 @@ private:
   std::vector<ASpreadSheetCell *> m_references;
 
 public:
-  uint32_t                 m_unresolvedRefs;
+  size_t                 m_unresolvedRefs;
 
 protected:
